ci: harden GitHub Actions workflows
Declare least-privilege permissions, pin third-party actions by SHA, disable checkout credential persistence, and remove cache usage from CI jobs.

Add a dedicated workflow-audit job that runs zizmor and move the CI workflow contract assertions into a focused test module so the hardening stays enforced.
